Style fix

+++ 0x18-dynamic_libraries/div.c
@@ -9,6 +9,6 @@ float div(int a, int b)
 {
 	float result = 0;
 
-	result = (float)a / b;
+	result = (float)(a) /(float)(b);
 	return (result);
 }
